pyproject.toml
[project]
name = "speech-emotion-recognition"
version = "1.0.0"
description = "Speech Emotion Recognition using CNN-LSTM for Mental Health Monitoring"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "tensorflow-macos>=2.14.0",
    "tensorflow-metal>=1.1.0",
    "librosa>=0.10.0",
    "numpy<2.0",
    "pandas",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "soundfile",
    "tqdm",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = []
